Remove `HalfEdge`'s `Reverse` implementation
